requirements.txt
streamlit
openai
streamlit-chat
python-dotenv
clarifai_grpc
protobuf~=3.20.0
altair==4.0
pymongo
